Added some things.
The player can now shoot and kill the centipede.
The player dies if he is hit by the centipede.  This is turned off right
now, because the player will continue to die from the centipede.
Added alternative to the movement.  It's very hard being restricted to
two lanes, and it doesn't feel right.  If other is still preferred, it's
a simple toggle from 1 to 0.

Bugs:
Sometimes the centipedes will go off screen still.
The centipede only collides with non damaged "mushrooms".
When a centipede is killed.  It doesn't spawn a "mushroom" in the
correct tile.  It's close, but it isn't.